requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
blinker==1.8.2
cattrs==24.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
Mako==1.3.5
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
openai==1.42.0
platformdirs==4.2.2
pydantic==2.8.2
pydantic_core==2.20.1
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
requests-cache==1.2.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
tqdm==4.66.5
typing_extensions==4.12.2
url-normalize==1.4.3
urllib3==2.2.2
Werkzeug==3.0.4
